feat: export firebase stuff into separate file and use a singleton
