+++ backend/web/middleware.go
@@ -5,27 +5,20 @@ import (
 	"fmt"
 	"log"
 	"net/http"
-	"os"
 	"time"
 
 	"cloud.google.com/go/firestore"
-	firebase "firebase.google.com/go"
 	"firebase.google.com/go/auth"
 	"github.com/labstack/echo/v4"
-	"google.golang.org/api/option"
 )
 
 var LastGlobalHealthCheck int = 0
 
 func FirestoreMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
-
 		ctx := context.Background()
-		app, err := InitFirebase()
 
-		if err != nil {
-			log.Fatalf("Error initializing Firebase app: %v", err)
-		}
+		app := GetFirebase()
 
 		firestore, err := app.Firestore(ctx)
 		if err != nil {
@@ -83,26 +76,6 @@ func checkGlobalHealth(firestore *firestore.Client, LastGlobalHealthCheck int) {
 		//
 		// }
 	}
-}
-
-func InitFirebase() (*firebase.App, error) {
-	ctx := context.Background()
-
-	location, exists := os.LookupEnv("FIRESTORE_SERVICE_ACCOUNT_PATH")
-
-	if !exists {
-		log.Fatalf("FIRESTORE_SERVICE_ACCOUNT_PATH not set in environment. \n")
-		return nil, fmt.Errorf("FIRESTORE_SERVICE_ACCOUNT_PATH not set in environment. \n")
-	}
-
-	opt := option.WithCredentialsFile(location)
-
-	app, err := firebase.NewApp(ctx, nil, opt)
-	if err != nil {
-		log.Fatalf("Error initializing Firebase app: %v", err)
-	}
-
-	return app, nil
 }
 
 func AuthMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
